Revert "WOR-1327 fix dashboard polling (#4426)"
This reverts commit cbfdab126700a1907d29c984eb17dd08cb6d3e37.
