+++ src/pages/workspaces/WorkspacesList/WorkspacesList.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash/fp';
-import { ReactNode, useMemo, useState } from 'react';
+import { ReactNode, useEffect, useMemo, useState } from 'react';
 import { div, h, p } from 'react-hyperscript-helpers';
 import { isAzureUser } from 'src/auth/auth';
 import { Link, topSpinnerOverlay, transparentSpinnerOverlay } from 'src/components/common';
@@ -10,10 +10,10 @@ import { Ajax } from 'src/libs/ajax';
 import { withErrorIgnoring } from 'src/libs/error';
 import { updateSearch, useRoute } from 'src/libs/nav';
 import { useOnMount } from 'src/libs/react-utils';
+import { workspacesStore, workspaceStore } from 'src/libs/state';
 import { elements as StyleElements } from 'src/libs/style';
-import { newTabLinkProps } from 'src/libs/utils';
-import { cloudProviderTypes, WorkspaceWrapper as Workspace } from 'src/libs/workspace-utils';
-import { useDeletionPolling } from 'src/pages/workspaces/hooks/useDeletionPolling';
+import { newTabLinkProps, pollWithCancellation } from 'src/libs/utils';
+import { cloudProviderTypes, WorkspaceState, WorkspaceWrapper as Workspace } from 'src/libs/workspace-utils';
 import { categorizeWorkspaces } from 'src/pages/workspaces/WorkspacesList/CategorizedWorkspaces';
 import { RecentlyViewedWorkspaces } from 'src/pages/workspaces/WorkspacesList/RecentlyViewedWorkspaces';
 import { useWorkspacesWithSubmissionStats } from 'src/pages/workspaces/WorkspacesList/useWorkspacesWithSubmissionStats';
@@ -43,7 +43,7 @@ export const WorkspacesList = (): ReactNode => {
   } = useWorkspacesWithSubmissionStats();
 
   const [featuredList, setFeaturedList] = useState<{ name: string; namespace: string }[]>();
-  useDeletionPolling(workspaces);
+  useDeletetionPolling(workspaces);
   const { query } = useRoute();
   const filters: WorkspaceFilterValues = getWorkspaceFiltersFromQuery(query);
 
@@ -110,4 +110,58 @@ export const WorkspacesList = (): ReactNode => {
       ]),
     ]),
   ]);
+};
+
+const useDeletetionPolling = (workspaces: Workspace[]) => {
+  // we have to do the signal/abort manually instead of with useCancelable so that the it can be cleaned up in the
+  // this component's useEffect, instead of the useEffect in useCancelable
+  const [controller, setController] = useState(new window.AbortController());
+  const abort = () => {
+    controller.abort();
+    setController(new window.AbortController());
+  };
+  useEffect(() => {
+    const updateWorkspacesStore = (workspace: Workspace, state: WorkspaceState, errorMessage?: string): Workspace[] => {
+      const updateList = _.cloneDeep(workspaces);
+      const updateWS = _.find(
+        (ws: Workspace) => ws.workspace.workspaceId === workspace.workspace.workspaceId,
+        updateList
+      )!;
+      updateWS.workspace.state = state;
+      updateWS.workspace.errorMessage = errorMessage;
+      return updateList;
+    };
+
+    const checkWorkspaceDeletion = async (workspace: Workspace) => {
+      try {
+        const wsResp: Workspace = await Ajax(controller.signal)
+          .Workspaces.workspace(workspace.workspace.namespace, workspace.workspace.name)
+          .details(['workspace.state', 'workspace.errorMessage']);
+        const state = wsResp.workspace.state;
+        if (state === 'DeleteFailed') {
+          abort();
+          workspacesStore.update(() => updateWorkspacesStore(workspace, state, wsResp.workspace.errorMessage));
+          workspaceStore.reset();
+        }
+      } catch (error) {
+        if (error instanceof Response && error.status === 404) {
+          abort();
+          workspacesStore.update(() => updateWorkspacesStore(workspace, 'Deleted'));
+          workspaceStore.reset();
+        }
+      }
+    };
+    const iterateDeletingWorkspaces = async () => {
+      const deletingWorkspaces = _.filter((ws) => ws.workspace.state === 'Deleting', workspaces);
+      for (const ws of deletingWorkspaces) {
+        await checkWorkspaceDeletion(ws);
+      }
+    };
+
+    pollWithCancellation(() => iterateDeletingWorkspaces(), 30000, false, controller.signal);
+    return () => {
+      abort();
+    };
+    //  eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [workspaces]); // adding the controller to deps causes a double fire of the effect
 };
